implement views folder structure
