+++ api/__init__.py
@@ -1,7 +1,10 @@
 from flask import Flask, Blueprint
 from api.v2.database import Database
-from api.v2 import v2
-from api.v1 import v1
+from api.v2.views.authentication import v2_authentication
+from api.v2.views.party import v2_party
+from api.v2.views.office import v2_offices
+from api.v1.views.office import v1_office
+from api.v1.views.party import v1_parties
 from api.v2.models.errors import ConfigError, InputError
 
 def create_app(configuration='config.Default'):
@@ -9,6 +12,9 @@ def create_app(configuration='config.Default'):
     app.config.from_object(configuration)
     database = Database(app.config.get("DATABASE"))
     database.init_db()
-    app.register_blueprint(v1)
-    app.register_blueprint(v2)
+    app.register_blueprint(v1_office)
+    app.register_blueprint(v1_parties)
+    app.register_blueprint(v2_authentication)
+    app.register_blueprint(v2_offices)
+    app.register_blueprint(v2_party)
     return app
